// src/app/models/modules/module-standards.ts
export class ModuleStandards {
  retriesMax: number;
  retriesUsed: number;
  invalidTryDelay: number;
  retriesReachedDelay: number;
  actionInvalidTry: string;
  actionRetriesReached: string;
  actionValid: string;
  isValidated: boolean;
  retriesReachedEnd: number;
  enabled: boolean;
  enableDelay: number;
  disableDelay: number;
  allowInteraction: boolean;
}